requirements.txt
langchain
langgraph
chromadb
faiss-cpu
streamlit
pypdf
python-docx
openai
huggingface_hub
tiktoken
reportlab 
